no eager loading by default
The description for get(), paginate() and simplepaginate() states that relations are eager loaded, but they are not. By default they are lazy loaded. The `with`  method call is what actually does eager loading. 

If its this simple, then removing the statement about eager loading is all is needed to make it clearer.